PDF和Markdown合并工具/requirements.txt
PyInstaller==6.11.1
Pillow==10.2.0
python-docx==1.1.0
PyMuPDF==1.23.26
markdown==3.5.2